requirements.txt
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
google-api-core==2.10.1
google-auth==2.11.1
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
idna==3.4
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
pymongo==4.2.0
requests==2.28.1
rsa==4.9
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.12
